refactor(data, domain): 매핑 로직 DTO 이동 및 Enum 처리 개선을 통한 SRP/OCP 강화
- Mapper 로직의 SRP 및 OCP 강화를 위한 구조 개선
- 불필요해진 구형 Mapper 파일들 제거 및 패키지 구조 최적화
- 각 DTO 파일 하단에 확장 함수(toDomain, toDto) 정의 및 관련 임포트 경로 업데이트
- Enum 클래스 내부에 from() 메서드를 추가하여 문자열 파싱 및 비즈니스 규칙 캡슐화
- 외부 API(Kakao) 응답 모델의 변환 로직을 해당 응답 DTO 파일로 이동하여 공통 DTO의 독립성 확보

Refs #81
