+++ data/src/main/java/com/moon/pharm/data/datasource/remote/dto/IntakeRecordDTO.kt
@@ -5,6 +5,8 @@ import com.google.firebase.firestore.DocumentId
 import com.google.firebase.firestore.IgnoreExtraProperties
 import com.google.firebase.firestore.PropertyName
 import com.moon.pharm.data.common.EMPTY_STRING
+import com.moon.pharm.data.common.toTimestamp
+import com.moon.pharm.domain.model.medication.IntakeRecord
 
 @IgnoreExtraProperties
 data class IntakeRecordDTO(
@@ -19,4 +21,24 @@ data class IntakeRecordDTO(
     @get:PropertyName("isTaken")
     val isTaken: Boolean = false,
     val takenTime: Timestamp? = null
+)
+
+// IntakeRecord 매퍼
+fun IntakeRecordDTO.toDomain(): IntakeRecord = IntakeRecord(
+    id = this.id,
+    userId = this.userId,
+    medicationId = this.medicationId,
+    scheduleId = this.scheduleId,
+    recordDate = this.recordDate,
+    isTaken = this.isTaken,
+    takenTime = this.takenTime?.toDate()?.time
+)
+
+fun IntakeRecord.toDto(): IntakeRecordDTO = IntakeRecordDTO(
+    userId = this.userId,
+    medicationId = this.medicationId,
+    scheduleId = this.scheduleId,
+    recordDate = this.recordDate,
+    isTaken = this.isTaken,
+    takenTime = this.takenTime?.toTimestamp()
 )
