+++ data/src/main/java/com/moon/pharm/data/repository/MedicationRepositoryImpl.kt
@@ -2,9 +2,9 @@ package com.moon.pharm.data.repository
 
 import com.google.firebase.firestore.FirebaseFirestoreException
 import com.moon.pharm.data.datasource.MedicationDataSource
+import com.moon.pharm.data.datasource.remote.dto.toDomain
+import com.moon.pharm.data.datasource.remote.dto.toDto
 import com.moon.pharm.data.di.IoDispatcher
-import com.moon.pharm.data.mapper.toDomain
-import com.moon.pharm.data.mapper.toDto
 import com.moon.pharm.domain.model.medication.IntakeRecord
 import com.moon.pharm.domain.model.medication.Medication
 import com.moon.pharm.domain.model.medication.MedicationException
