+++ data/src/main/java/com/moon/pharm/data/repository/PharmacyRepositoryImpl.kt
@@ -3,9 +3,9 @@ package com.moon.pharm.data.repository
 import com.google.firebase.firestore.FirebaseFirestoreException
 import com.moon.pharm.data.datasource.PharmacySearchDataSource
 import com.moon.pharm.data.datasource.PharmacyStorageDataSource
+import com.moon.pharm.data.datasource.remote.dto.toDomain
+import com.moon.pharm.data.datasource.remote.dto.toDto
 import com.moon.pharm.data.di.IoDispatcher
-import com.moon.pharm.data.mapper.toDomain
-import com.moon.pharm.data.mapper.toDto
 import com.moon.pharm.domain.model.pharmacy.Pharmacy
 import com.moon.pharm.domain.model.pharmacy.PharmacyException
 import com.moon.pharm.domain.repository.PharmacyRepository
