+++ domain/src/main/java/com/moon/pharm/domain/model/auth/UserType.kt
@@ -1,5 +1,11 @@
 package com.moon.pharm.domain.model.auth
 
 enum class UserType {
-    GENERAL, PHARMACIST
+    GENERAL, PHARMACIST;
+
+    companion object {
+        fun from(value: String?): UserType {
+            return entries.find { it.name == value } ?: GENERAL
+        }
+    }
 }
