+++ domain/src/main/java/com/moon/pharm/domain/model/medication/MealTiming.kt
@@ -4,5 +4,11 @@ enum class MealTiming(val label: String) {
     BEFORE_MEAL("식전"),
     DURING_MEAL("식사 중"),
     AFTER_MEAL("식후"),
-    NONE("상관없음")
+    NONE("상관없음");
+
+    companion object {
+        fun from(value: String?): MealTiming {
+            return entries.find { it.name == value } ?: NONE
+        }
+    }
 }
