+++ domain/src/main/java/com/moon/pharm/domain/model/medication/MedicationType.kt
@@ -3,5 +3,11 @@ package com.moon.pharm.domain.model.medication
 enum class MedicationType(val label: String) {
     PRESCRIPTION("처방약"),
     OTC("일반약"),
-    SUPPLEMENT("비타민/영양제")
+    SUPPLEMENT("비타민/영양제");
+
+    companion object {
+        fun from(value: String?): MedicationType {
+            return entries.find { it.name == value } ?: OTC
+        }
+    }
 }
